backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
chromadb
pypdf
tiktoken
python-multipart
